Apply changes for V2

+++ app/src/main/java/com/example/scrollbooker/entity/user/userProfile/data/remote/UserProfileDto.kt
@@ -100,5 +100,5 @@ data class UserCountersDto(
     val ratingsCount: Int,
 
     @SerializedName("ratings_average")
-    val ratingsAverage: BigDecimal
+    val ratingsAverage: Float
 )
